chore(ci): update deps and fix frontend quality
