+++ backend/pipelines/bbr_buildings/pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
     "python-dotenv>=1.0.0",
     "pyarrow>=14.0.0",
     "s3fs",
-    "lxml>=4.9.0",
+    "lxml>=6.0.4,<6.1",
     "tqdm>=4.67.3",
     "landbruget-common",
 ]
@@ -41,7 +41,7 @@ bbr-buildings = "bbr_buildings_pipeline.main:main"
 
 [dependency-groups]
 dev = [
-    "pytest>=8.0.0",
+    "pytest>=9.0.3",
     "pytest-cov>=4.0.0",
     "ruff",
     "ty",
