+++ backend/pipelines/bmd_scraper/pyproject.toml
@@ -33,7 +33,7 @@ packages = ["bronze", "silver"]
 
 [dependency-groups]
 dev = [
-    "pytest>=8.0.0",
+    "pytest>=9.0.3",
     "pytest-cov>=4.0.0",
     "ruff",
     "ty",
