+++ backend/pipelines/dma_scraper/pyproject.toml
@@ -35,7 +35,7 @@ packages = ["bronze", "silver"]
 
 [dependency-groups]
 dev = [
-    "pytest>=8.0.0",
+    "pytest>=9.0.3",
     "pytest-cov>=4.0.0",
     "ruff",
     "ty",
