+++ backend/pipelines/h3_pfas_exposure_pipeline/pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
     "s3fs",
     # Utilities
     "simple-singleton>=2.0.0",
-    "psutil>=5.9.0",
+    "psutil>=7.2.2",
     "tqdm>=4.66.0",
     "landbruget-common",
 ]
@@ -37,7 +37,7 @@ packages = ["src/h3_pfas_exposure"]
 
 [dependency-groups]
 dev = [
-    "pytest>=8.0.0",
+    "pytest>=9.0.3",
     "pytest-asyncio>=0.21.0",
     "pytest-cov>=4.0.0",
     "ruff",
