+++ backend/pipelines/svineflytning_pipeline/pyproject.toml
@@ -11,11 +11,11 @@ authors = [
 dependencies = [
     "zeep~=4.3.1",
     "certifi~=2026.2.25",
-    "lxml~=5.3.0",
+    "lxml~=6.0.4",
     "xmlsec==1.3.17",
     "s3fs",
     "python-dotenv~=1.1.1",
-    "cryptography>=44.0.0",
+    "cryptography>=46.0.7",
     "tqdm~=4.67.1",
     "xmltodict~=1.0.4",
     "tenacity>=9.1.2",
@@ -38,7 +38,7 @@ packages = ["bronze", "silver"]
 
 [dependency-groups]
 dev = [
-    "pytest>=8.0.0",
+    "pytest>=9.0.3",
     "pytest-cov>=4.0.0",
     "ruff",
     "ty",
