+++ frontend/src/components/pesticidkort/FieldList.tsx
@@ -1,70 +1,14 @@
 'use client';
 
-import { useMemo, useState, useEffect, useRef } from 'react';
+import { useMemo, useRef, useState } from 'react';
 import { ChevronDown } from 'lucide-react';
 import { motion, useReducedMotion } from 'motion/react';
 import { FieldCard } from '@/components/pesticidkort/FieldCard';
 import type { HistogramBin } from '@/components/pesticidkort/BurdenScale';
 import type { NearbyFieldSummary } from '@/components/pesticidkort/types';
+import { useSelectedFieldScroll } from '@/components/pesticidkort/useSelectedFieldScroll';
 
 const INITIAL_VISIBLE = 5;
-const CARD_SCROLL_PADDING = 12;
-const EXPANSION_OBSERVER_WINDOW_MS = 400;
-
-function findScrollContainer(element: HTMLElement): HTMLElement | null {
-  let current = element.parentElement;
-
-  while (current) {
-    const style = window.getComputedStyle(current);
-    const overflowY = style.overflowY;
-    if (
-      (overflowY === 'auto' || overflowY === 'scroll') &&
-      current.scrollHeight > current.clientHeight
-    ) {
-      return current;
-    }
-    current = current.parentElement;
-  }
-
-  return document.scrollingElement instanceof HTMLElement
-    ? document.scrollingElement
-    : null;
-}
-
-function ensureCardFullyVisible(
-  element: HTMLElement,
-  behavior: ScrollBehavior = 'smooth'
-) {
-  const container = findScrollContainer(element);
-
-  if (!container) {
-    element.scrollIntoView({ behavior, block: 'nearest' });
-    return;
-  }
-
-  const containerRect =
-    container === document.scrollingElement
-      ? { top: 0, bottom: window.innerHeight }
-      : container.getBoundingClientRect();
-  const elementRect = element.getBoundingClientRect();
-  const visibleTop = containerRect.top + CARD_SCROLL_PADDING;
-  const visibleBottom = containerRect.bottom - CARD_SCROLL_PADDING;
-
-  if (elementRect.top < visibleTop) {
-    container.scrollBy({
-      top: elementRect.top - visibleTop,
-      behavior,
-    });
-    return;
-  }
-
-  if (elementRect.bottom > visibleBottom) {
-    container.scrollBy({
-      top: elementRect.bottom - visibleBottom,
-      behavior,
-    });
-  }
-}
 
 interface FieldListProps {
   fields: NearbyFieldSummary[];
@@ -89,6 +33,10 @@ export function FieldList({
     () => [...fields].sort((a, b) => a.distance_m - b.distance_m),
     [fields]
   );
+  const sortedFieldIds = useMemo(
+    () => sortedFields.map((field) => field.field_uuid),
+    [sortedFields]
+  );
 
   const visibleFields = showAll
     ? sortedFields
@@ -100,60 +48,14 @@ export function FieldList({
     selectedFieldUuid === clickedField.field_uuid &&
     !sortedFields.some((f) => f.field_uuid === selectedFieldUuid);
 
-  useEffect(() => {
-    if (!selectedFieldUuid) return;
-
-    let animationFrame1 = 0;
-    let animationFrame2 = 0;
-    let resizeObserver: ResizeObserver | null = null;
-    let resizeObserverTimeout = 0;
-
-    const scrollSelectedCard = (behavior: ScrollBehavior = 'smooth') => {
-      const selectedCard = cardRefs.current.get(selectedFieldUuid);
-      if (!selectedCard) return;
-
-      ensureCardFullyVisible(selectedCard, behavior);
-
-      if (typeof ResizeObserver === 'undefined') return;
-
-      resizeObserver?.disconnect();
-      resizeObserver = new ResizeObserver(() => {
-        ensureCardFullyVisible(selectedCard, 'auto');
-      });
-      resizeObserver.observe(selectedCard);
-
-      window.clearTimeout(resizeObserverTimeout);
-      resizeObserverTimeout = window.setTimeout(() => {
-        resizeObserver?.disconnect();
-        resizeObserver = null;
-      }, EXPANSION_OBSERVER_WINDOW_MS);
-    };
-
-    const scheduleScroll = () => {
-      animationFrame1 = requestAnimationFrame(() => {
-        animationFrame2 = requestAnimationFrame(() => {
-          scrollSelectedCard();
-        });
-      });
-    };
-
-    const idx = sortedFields.findIndex(
-      (f) => f.field_uuid === selectedFieldUuid
-    );
-    if (idx >= INITIAL_VISIBLE && !showAll) {
-      setShowAll(true);
-      scheduleScroll();
-    } else {
-      scheduleScroll();
-    }
-
-    return () => {
-      cancelAnimationFrame(animationFrame1);
-      cancelAnimationFrame(animationFrame2);
-      resizeObserver?.disconnect();
-      window.clearTimeout(resizeObserverTimeout);
-    };
-  }, [selectedFieldUuid, sortedFields, showAll]);
+  useSelectedFieldScroll({
+    selectedFieldUuid,
+    sortedFieldIds,
+    showAll,
+    setShowAll,
+    cardRefs,
+    initialVisible: INITIAL_VISIBLE,
+  });
 
   return (
     <div>
